Remove named exports from pages
