+++ site/gatsby-site/pages/apps/discover.js
@@ -1,22 +1,17 @@
 import React, { useState, useEffect, useRef, useMemo, useCallback } from 'react';
-import { navigate } from 'gatsby';
 import { StringParam, QueryParams, useQueryParams } from 'use-query-params';
 import algoliasearch from 'algoliasearch/lite';
 import { debounce } from 'debounce';
 import {
   InstantSearch,
   Highlight,
-  connectHits,
   connectRefinementList,
   Pagination,
   Stats,
   connectSearchBox,
-  connectStateResults,
   connectRange,
 } from 'react-instantsearch-dom';
 import styled from 'styled-components';
-import config from '../../config';
-import md5 from 'md5';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import {
   faNewspaper,
@@ -27,18 +22,18 @@ import {
   faTimesCircle,
   faCalendarAlt,
 } from '@fortawesome/free-solid-svg-icons';
-import { useModal, CustomModal } from '../../src/components/useModal';
 import LayoutHideSidebar from 'components/LayoutHideSidebar';
-import { Form, Button, Spinner } from 'react-bootstrap';
+import { Form, Button } from 'react-bootstrap';
 import Helmet from 'react-helmet';
 
 import '../../static/discover/src/app.css';
 import '../../static/discover/src/index.css';
 import { add, format, formatISO, isAfter, isBefore } from 'date-fns';
+import { Hits } from '../../src/components/IncidentCards';
 
-export const indexName = 'instant_search';
+const indexName = 'instant_search';
 
-export const searchClient = algoliasearch('JD5JCVZEVS', 'c5e99d93261645721a1765fe4414389c');
+const searchClient = algoliasearch('JD5JCVZEVS', 'c5e99d93261645721a1765fe4414389c');
 
 //// Alternative
 // const indexName = "aiid-emergency";
@@ -165,13 +160,6 @@ const RangeInputContainer = styled.div`
   padding-bottom: 1.5rem;
 `;
 
-const StatsContainer = styled.div`
-  display: grid;
-  max-width: 100%;
-  grid-template-columns: 1fr 3fr;
-  padding: 1.25rem;
-`;
-
 const FacetsSide = styled.div`
   display: flex;
   flex-direction: column;
@@ -184,10 +172,6 @@ const FacetsSide = styled.div`
   @media (max-width: 570px) {
     max-width: unset;
   }
-`;
-
-const Text = styled.p`
-  line-height: 1.3;
 `;
 
 const ResultsSide = styled.div`
@@ -205,14 +189,6 @@ const ResultsSide = styled.div`
     width: 100%;
     padding: 0;
   }
-`;
-
-const IncidentCardContainer = styled.div`
-  border: 1.5px solid #d9deee;
-  border-radius: 5px;
-  box-shadow: 0 2px 5px 0px #e3e5ec;
-  display: flex;
-  flex-direction: column;
 `;
 
 const RefinementListContainer = styled.div`
@@ -298,14 +274,6 @@ const StyledPagination = styled(Pagination)`
   }
 `;
 
-const CardFooter = styled.div`
-  display: flex;
-  width: 100%;
-  flex-direction: row;
-  justify-content: space-around;
-  align-items: center;
-`;
-
 const RefinementListHeader = styled.span`
   display: block;
 `;
@@ -347,10 +315,6 @@ const StyledButton = styled.button`
     background-color: #007bff;
     border-color: #007bff;
   `};
-`;
-
-const CardBody = styled.div`
-  padding: 1.25rem 1.25rem 0 1.25rem !important;
 `;
 
 const StyledSearchInput = styled.input`
@@ -434,25 +398,6 @@ const Header = styled.div`
   }
 `;
 
-export const NoResults = styled.div`
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-
-  grid-column-start: 2;
-  grid-column-end: 3;
-
-  p {
-    text-align: center;
-  }
-
-  @media (max-width: 1240px) {
-    grid-column-start: 1;
-  }
-`;
-
 const SearchContainer = styled.div`
   flex-grow: 1;
 `;
@@ -460,10 +405,6 @@ const SearchContainer = styled.div`
 const SearchForm = styled.form`
   display: block;
   position: relative;
-`;
-
-const StyledLabel = styled.p`
-  margin: 0.6em 0;
 `;
 
 const StyledRefinementList = ({
@@ -504,226 +445,6 @@ const StyledRefinementList = ({
       onChange={(event) => searchForItems(event.currentTarget.value)}
     />
   </RefinementListContainer>
-);
-
-const getFlagModalContent = () => (
-  <div className="modal-body">
-    <p>Is there a problem with this content? Examples of &quot;problems`&quot;` include,</p>
-    <ul>
-      <li>The text contents of the report are not parsed properly</li>
-      <li>The authors of the report are not associated with the report</li>
-      <li>The report is associated with the wrong incident</li>
-      <li>The text contents of the report are not parsed properly</li>
-    </ul>
-    <p>
-      Flagged content will still be displayed within the database, but a database editor will
-      periodically review the incident reports that have been flagged. Please note that the content
-      contained within incident reports (e.g., the commentary within a news article) does not need
-      to be correct or consistent across articles. If an article is wrong, misleading, or
-      fraudulent, the best response is to submit additional incident reports that correct the
-      record. The incident database is meant to capture the complete state of knowledge and
-      discourse for incidents, not as an arbiter of what happened in individual incidents. In future
-      versions of the database it will additionally be possible to apply tags to incident reports to
-      classNameify their content.
-    </p>
-    <p>Please do NOT flag content if,</p>
-    <ul>
-      <li>You disagree with the report</li>
-      <li>The linked report has disappeared from the web</li>
-      <li>The report should not be considered an `&quot;`incident`&quot;`</li>
-    </ul>
-    <button
-      type="button"
-      className="btn btn-danger btn-sm btn-block flag-button"
-      data-dismiss="modal"
-    >
-      Flag Report
-    </button>
-  </div>
-);
-
-const cardNeedsBlockquote = (item) => {
-  if (item.text && item.text.matchLevel === 'full') {
-    return true;
-  }
-  return false;
-};
-
-const getParagraphs = (itemText) => {
-  return (
-    <>
-      {itemText.split('\n').map((paragraph, index, array) => (
-        <p key={index}>
-          {array.length - 1 === index ? <>{paragraph + '...'}</> : <>{paragraph}</>}
-        </p>
-      ))}
-    </>
-  );
-};
-
-export const IncidentStatsCard = ({ incidentId, reportCount, incidentDate }) => {
-  const STATS = [
-    {
-      key: 'incidentId',
-      label: 'Incident ID',
-    },
-    {
-      key: 'reportCount',
-      label: 'Report Count',
-    },
-    {
-      key: 'incidentDate',
-      label: 'Incident Date',
-    },
-  ];
-
-  if (reportCount === 0) {
-    return null;
-  }
-
-  const stats = {
-    incidentId,
-    reportCount,
-    incidentDate,
-  };
-
-  return (
-    <IncidentCardContainer className="card">
-      <div className="card-header">
-        <h4>Incident Stats</h4>
-      </div>
-      <StatsContainer className="card-body">
-        <div>
-          {STATS.map((stat) => (
-            <div key={stat.key}>{stat.label}</div>
-          ))}
-        </div>
-        <div>
-          {STATS.map((stat) => (
-            <div key={stat.key}>{stats[stat.key]}</div>
-          ))}
-        </div>
-      </StatsContainer>
-    </IncidentCardContainer>
-  );
-};
-
-export const IncidentCard = ({
-  item,
-  authorsModal,
-  submittersModal,
-  flagReportModal,
-  toggleFilterByIncidentId,
-  showDetails,
-  isCitePage,
-}) => (
-  <IncidentCardContainer id={item.mongodb_id}>
-    <div className="card-header">
-      {isCitePage ? <span>{item.title}</span> : <Highlight hit={item} attribute="title" />}
-      <p className="subhead">
-        {item.source_domain} &middot;{' '}
-        {item.date_published ? item.date_published.substring(0, 4) : 'Needs publish date'}
-      </p>
-    </div>
-    <CardBody className="card-body">
-      {/* <Highlight hit={item} attribute="description" /> */}
-      {item._snippetResult && cardNeedsBlockquote(item._snippetResult) && (
-        <blockquote>
-          <Highlight
-            hit={{
-              _highlightResult: {
-                ...item._snippetResult,
-                text: {
-                  ...item._snippetResult.text,
-                  value: `...${item._snippetResult.text.value}...`,
-                },
-              },
-            }}
-            attribute="text"
-          />
-        </blockquote>
-      )}
-      {showDetails ? (
-        <div>{getParagraphs(item.text)}</div>
-      ) : (
-        <div>
-          <Text>{item.text.substr(0, 400) + '...'}</Text>
-        </div>
-      )}
-    </CardBody>
-    <div className="align-bottom">
-      <img className="image-preview" alt={item.title} src={getImageHashPath(item.image_url)} />
-      {toggleFilterByIncidentId && (
-        <button
-          type="button"
-          className="btn btn-secondary btn-sm btn-block assignment-button"
-          onClick={() => {
-            navigate(`/cite/${item.incident_id}#${item.mongodb_id}`);
-          }}
-        >
-          <StyledLabel>Show Details on Incident #{item.incident_id}</StyledLabel>
-        </button>
-      )}
-    </div>
-    <CardFooter className="card-footer text-muted">
-      <a href={item.url} target="_blank" rel="noreferrer">
-        <FontAwesomeIcon icon={faNewspaper} className="far fa-newspaper" title="Read the Source" />
-      </a>
-
-      <FontAwesomeIcon
-        icon={faIdCard}
-        className="pointer far fa-id-card"
-        title="Authors"
-        onClick={() =>
-          authorsModal.openFor({
-            title: 'Authors',
-            body: function f() {
-              return item.authors.join(', ');
-            },
-          })
-        }
-      />
-
-      <FontAwesomeIcon
-        icon={faUserShield}
-        className="pointer fas fa-user-shield"
-        title="Submitters"
-        onClick={() =>
-          submittersModal.openFor({
-            title: 'Submitters',
-            body: function f() {
-              return item.submitters.join(', ');
-            },
-          })
-        }
-      />
-
-      <FontAwesomeIcon
-        icon={faFlag}
-        className="pointer far fa-flag"
-        title="Flag Report"
-        onClick={() =>
-          flagReportModal.openFor({
-            title: 'Submitters',
-            body: function f() {
-              return getFlagModalContent();
-            },
-          })
-        }
-      />
-
-      {toggleFilterByIncidentId && (
-        <span
-          aria-hidden="true"
-          className="pointer"
-          onClick={() => toggleFilterByIncidentId(item.incident_id + '')}
-        >
-          <FontAwesomeIcon icon={faHashtag} className="fas fa-hashtag" title="Incident ID" />
-          {item.incident_id}
-        </span>
-      )}
-    </CardFooter>
-  </IncidentCardContainer>
 );
 
 // eslint-disable-next-line react/display-name
@@ -787,10 +508,6 @@ const StyledSearchBox = React.memo(({ refine, customRef }) => {
 const CustomSearchBox = React.memo(connectSearchBox(StyledSearchBox));
 
 const RefinementList = connectRefinementList(StyledRefinementList);
-
-const getImageHashPath = (imgUrl) => {
-  return `${config.gatsby.siteUrl}/large_media/report_banners/${md5(imgUrl)}`;
-};
 
 const removeUndefinedAttributes = (obj) => {
   for (const attr in obj) {
@@ -996,53 +713,6 @@ const RangeInput = ({ currentRefinement: { min, max }, refine }) => {
 
 const CustomRangeInput = connectRange(RangeInput);
 
-const RenderCards = ({
-  hits,
-  toggleFilterByIncidentId,
-  showDetails,
-  authorsModal,
-  submittersModal,
-  flagReportModal,
-  isSearchStalled,
-}) => {
-  if (isSearchStalled) {
-    return (
-      <NoResults>
-        <Spinner animation="border" role="status" variant="primary">
-          <span className="sr-only">Loading...</span>
-        </Spinner>
-      </NoResults>
-    );
-  }
-
-  if (hits.length === 0) {
-    return (
-      <NoResults>
-        <p>Your search returned no results.</p>
-        <p>Please clear your search in the search box above or the filters.</p>
-      </NoResults>
-    );
-  }
-
-  return (
-    <>
-      {hits.map((hit) => (
-        <IncidentCard
-          key={hit.objectID}
-          item={hit}
-          authorsModal={authorsModal}
-          submittersModal={submittersModal}
-          flagReportModal={flagReportModal}
-          toggleFilterByIncidentId={toggleFilterByIncidentId}
-          showDetails={showDetails}
-        />
-      ))}
-    </>
-  );
-};
-
-const CustomHits = connectHits(connectStateResults(RenderCards));
-
 const FiltersBar = ({ filters, updateFilters, updateQuery }) => {
   const flitersArray = [];
 
@@ -1137,36 +807,6 @@ const FiltersBar = ({ filters, updateFilters, updateQuery }) => {
         </Tags>
       ))}
     </FiltersContainer>
-  );
-};
-
-export const Hits = ({
-  toggleFilterByIncidentId,
-  showDetails = false,
-  sortByDatePublished,
-  scrollTo,
-}) => {
-  const authorsModal = useModal();
-
-  const submittersModal = useModal();
-
-  const flagReportModal = useModal();
-
-  return (
-    <>
-      <CustomHits
-        toggleFilterByIncidentId={toggleFilterByIncidentId}
-        authorsModal={authorsModal}
-        submittersModal={submittersModal}
-        flagReportModal={flagReportModal}
-        showDetails={showDetails}
-        scrollTo={scrollTo}
-        sortByDatePublished={sortByDatePublished}
-      />
-      <CustomModal {...authorsModal} />
-      <CustomModal {...submittersModal} />
-      <CustomModal {...flagReportModal} />
-    </>
   );
 };
